feat: Expose proxy hosted zone id
This allows more complicated aliasing that currently possible with just the name exposed. Having both the domain name and hosted zone id will allow us to punch it into the alias on our own external `aws_route53_record` for a more custom approach to custom domains. The main reason for this is that we currently have different providers in different regions for our infrastructure and managing dns/route53, which as the providers differ means we can't use the currently available approach to custom domains.